Import version v2.11.1

+++ src/veda_context.cpp
@@ -2,6 +2,13 @@
 
 //------------------------------------------------------------------------------
 extern "C" {
+// implementation of VEDA API functions
+/**
+ * \defgroup vedaapi VEDA API
+ *
+ * To use VEDA API functions, include "veda.h" header.
+ */
+/** @{ */
 //------------------------------------------------------------------------------
 /**
  * @brief Destroy a VEDA context.
@@ -9,7 +16,8 @@ extern "C" {
  * @retval VEDA_SUCCESS on Success
  * @retval VEDA_ERROR_NOT_INITIALIZED VEDA library not initialized
  * @retval VEDA_ERROR_INVALID_DEVICE VEDA device id is not valid.
- * @retval VEDA_ERROR_UNKNOWN_CONTEXT VEDA context is not set for the calling thread.
+ * @retval VEDA_ERROR_UNKNOWN_CONTEXT VEDA context is not set for the calling thread.\n 
+ *
  * Destroys the VEDA context specified by ctx. The context ctx will be destroyed
  * regardless of how many threads it is current to. It is the responsibility of
  * the calling function to ensure that no API call issues using ctx while 
@@ -25,12 +33,14 @@ extern "C" {
  */
 VEDAresult	vedaCtxDestroy		(VEDAcontext ctx) {
 	GUARDED(
+	   if(ctx->device().ctx()){
 		uint32_t count = veda::Contexts::get_ctx_ins(ctx);
 		veda::Contexts::remove(ctx);
 		ctx->dec_ref_count(count);
 		ctx->destroy_proc_handle();
 		if(ctx->check_ref_count() == 0)
 			ctx->device().destroyCtx();
+	   }
 	)					
 }
 //------------------------------------------------------------------------------
@@ -51,7 +61,8 @@ VEDAresult	vedaCtxGetApiVersion	(VEDAcontext ctx, uint32_t* version)	{	GUARDED(*
  * @retval VEDA_ERROR_NOT_INITIALIZED VEDA library not initialized
  * @retval VEDA_ERROR_INVALID_DEVICE VEDA device id is not valid.
  * @retval VEDA_ERROR_UNKNOWN_CONTEXT VEDA context is not set for the calling thread.
- * @retval VEDA_ERROR_CONTEXT_IS_DESTROYED VEDA current context is already destroyed.
+ * @retval VEDA_ERROR_CONTEXT_IS_DESTROYED VEDA current context is already destroyed.\n 
+ *
  * Returns in *pctx the VEDA context bound to the calling CPU thread. 
  * If no context is bound to the calling CPU thread then *pctx is set to NULL
  * and VEDA_SUCCESS is returned.
@@ -65,7 +76,8 @@ VEDAresult	vedaCtxGetCurrent	(VEDAcontext* pctx)			{	GUARDED(*pctx = veda::Conte
  * @retval VEDA_ERROR_NOT_INITIALIZED VEDA library not initialized
  * @retval VEDA_ERROR_INVALID_DEVICE VEDA device id is not valid.
  * @retval VEDA_ERROR_UNKNOWN_CONTEXT VEDA context is not set for the calling thread.
- * @retval VEDA_ERROR_CONTEXT_IS_DESTROYED VEDA current context is already destroyed.
+ * @retval VEDA_ERROR_CONTEXT_IS_DESTROYED VEDA current context is already destroyed.\n 
+ *
  * Returns in *device the ordinal of the current context's device.
  */
 VEDAresult	vedaCtxGetDevice	(VEDAdevice* device)			{	GUARDED(*device = veda::Contexts::current()->device().vedaId();)	}
@@ -77,7 +89,8 @@ VEDAresult	vedaCtxGetDevice	(VEDAdevice* device)			{	GUARDED(*device = veda::Con
  * @retval VEDA_ERROR_NOT_INITIALIZED VEDA library not initialized
  * @retval VEDA_ERROR_INVALID_DEVICE VEDA device id is not valid.
  * @retval VEDA_ERROR_UNKNOWN_CONTEXT VEDA context is not set for the calling thread.
- * @retval VEDA_ERROR_CONTEXT_IS_DESTROYED VEDA current context is already destroyed.
+ * @retval VEDA_ERROR_CONTEXT_IS_DESTROYED VEDA current context is already destroyed.\n 
+ *
  * Pops the current VEDA context from the CPU thread and passes back the old
  * context handle in *pctx. That context may then be made current to a different
  * CPU thread by calling vedaCtxPushCurrent().
@@ -93,7 +106,8 @@ VEDAresult	vedaCtxPopCurrent	(VEDAcontext* pctx)			{	GUARDED(*pctx = veda::Conte
  * @param ctx Context to push.
  * @retval VEDA_SUCCESS on Success
  * @retval VEDA_ERROR_NOT_INITIALIZED VEDA library not initialized
- * @retval VEDA_ERROR_INVALID_DEVICE VEDA device id is not valid.
+ * @retval VEDA_ERROR_INVALID_DEVICE VEDA device id is not valid.\n 
+ *
  * Pushes the given context ctx onto the CPU thread's stack of current contexts.
  * The specified context becomes the CPU thread's current context, so all VEDA
  * functions that operate on the current context are affected. The previous
@@ -107,7 +121,8 @@ VEDAresult	vedaCtxPushCurrent	(VEDAcontext ctx)			{	GUARDED(veda::Contexts::push
  * @param ctx Context to bind to the calling CPU thread.
  * @retval VEDA_SUCCESS on Success
  * @retval VEDA_ERROR_NOT_INITIALIZED VEDA library not initialized
- * @retval VEDA_ERROR_INVALID_DEVICE VEDA device id is not valid.
+ * @retval VEDA_ERROR_INVALID_DEVICE VEDA device id is not valid.\n 
+ *
  * Binds the specified VEDA context to the calling CPU thread. If ctx is NULL 
  * then the VEDA context previously bound to the calling CPU thread is unbound
  * and VEDA_SUCCESS is returned.
@@ -126,7 +141,8 @@ VEDAresult	vedaCtxSetCurrent	(VEDAcontext ctx)			{	GUARDED(veda::Contexts::set(c
  * @retval VEDA_ERROR_NOT_INITIALIZED VEDA library not initialized
  * @retval VEDA_ERROR_INVALID_DEVICE VEDA device id is not valid.
  * @retval VEDA_ERROR_UNKNOWN_CONTEXT VEDA context is not set for the calling thread.
- * @retval VEDA_ERROR_CONTEXT_IS_DESTROYED VEDA current context is already destroyed.
+ * @retval VEDA_ERROR_CONTEXT_IS_DESTROYED VEDA current context is already destroyed.\n 
+ *
  * Returns the VEDA SM count of the VEDA context bound to the calling CPU thread.
  */
 VEDAresult	vedaCtxStreamCnt	(int* cnt)				{	GUARDED(*cnt = veda::Contexts::current()->streamCount();)		}
@@ -137,7 +153,8 @@ VEDAresult	vedaCtxStreamCnt	(int* cnt)				{	GUARDED(*cnt = veda::Contexts::curre
  * @retval VEDA_ERROR_NOT_INITIALIZED VEDA library not initialized
  * @retval VEDA_ERROR_INVALID_DEVICE VEDA device id is not valid.
  * @retval VEDA_ERROR_UNKNOWN_CONTEXT VEDA context is not set for the calling thread.
- * @retval VEDA_ERROR_CONTEXT_IS_DESTROYED VEDA current context is already destroyed.
+ * @retval VEDA_ERROR_CONTEXT_IS_DESTROYED VEDA current context is already destroyed.\n 
+ *
  * Blocks until the device has completed all preceding requested tasks.
  * vedaCtxSynchronize() returns an error if one of the preceding tasks failed.
  */
@@ -147,24 +164,27 @@ VEDAresult	vedaCtxSynchronize	(void)					{	GUARDED(veda::Contexts::current()->wr
 /**
  * @brief Create a VEDA context.
  * @param pctx Returned context handle of the new context.
- * @param flags Context creation flags
+ * @param mode Context mode may contain below two values –\n 
+ * VEDA_CONTEXT_MODE_OMP\n 
+ * VEDA_CONTEXT_MODE_SCALAR
  * @param dev Device to create context on
  * @retval VEDA_SUCCESS on Success
  * @retval VEDA_ERROR_NOT_INITIALIZED VEDA library not initialized
  * @retval VEDA_ERROR_INVALID_DEVICE VEDA device id is not valid.
  * @retval VEDA_ERROR_CANNOT_CREATE_CONTEXT Error while creating VEDA context.
- * @retval VEDA_ERROR_CANNOT_CREATE_STREAM error while creating VEDA SM.
+ * @retval VEDA_ERROR_CANNOT_CREATE_STREAM error while creating VEDA SM.\n 
+ *
  * Creates a new VEDA context and associates it with the calling thread. The 
  * flags parameter is described below. The caller of vedaCtxCreate() must call
  * vedaCtxDestroy() or when done using the context. If any context is already
  * current to the thread, it is changed by the newly created context and it may
- * be restored by a subsequent call to vedaCtxPopCurrent().
- * VEDA Context may be created with below two flags-
+ * be restored by a subsequent call to vedaCtxPopCurrent().\n 
+ * VEDA Context may be created with below two flags-\n 
  * VEDA_CONTEXT_MODE_OMP: In this mode, number of streaming multiprocessor(SM)
  * is always 1. However, in this mode the number of OPENMP threads in each 
  * VEDA SM is configured as the minimum value between the number of cores of 
  * the VEDA device and the user configured value for the number of OPENMP 
- * threads on VE device (set through environment variable).
+ * threads on VE device (set through environment variable).\n 
  * VEDA_CONTEXT_MODE_SCALAR: : In this mode number of streaming multiprocessor(SM)
  * are the minimum value between the number of cores of the VEDA device and the
  * user configured value for the number of OPENMP threads on VE device
@@ -177,6 +197,6 @@ VEDAresult vedaCtxCreate(VEDAcontext* pctx, int mode, VEDAdevice dev) {
 		veda::Contexts::push(*pctx);
 	)
 }
-
+/** @} */
 //------------------------------------------------------------------------------
 } // extern "C"
